Get CI testing working again and prevent future bit-rot (#165)
* Apply cargo fmt.

* Drop use of unaligned_references feature and clean up warnings.

* Drop "packed" annotation from Process.

It doesn't look like it's required, and it's causing unaligned
references that newer rust hates.

* Fix pasteos for syscall/interrupt stack alloc errors.

* Give the KERNEL/USER bit fields actual bit field values.

Future rust clippy errors about them doing "& 0" operations.  We can
decide if they're useful flags later if we actually want.

* Add a note about the apparently-dead kernel_stack.

Quiets clippy warnings in the strict-lint target.

* Drop slice_internals feature -- causes a warning, doesn't seem necessary.

* Drop redundant pub(self) annotations.

Fixes future clippy lints.

* Clean up some derefs that the compiler performs anyway.

Fixes future clippy lints.

* Drop redundant casts.

Fixes future clippy lints.

* Suppress or work around the new clippy 0-bitmask lints.

* Bump the nightly toolchain a bit so CI can have some hope again.

Fixes the new clippy complaints in the process.

* Pin the cargo-binutils and rustfilt versions in CI so it's more reproducible.

* Work around testing image build failure.

The RUN was throwing errors with current docker, but we can easily
enough just make it another ADD.

+++ runtime/x64/paging.rs
@@ -91,7 +91,7 @@ fn duplicate_table(original_table_paddr: PAddr, level: usize) -> Result<PAddr, P
 
     debug_assert!(level > 0);
     for i in 0..ENTRIES_PER_TABLE {
-        let entry = unsafe { *orig_table.offset(i as isize) };
+        let entry = unsafe { *orig_table.offset(i) };
         let paddr = entry_paddr(entry);
 
         // Check if we need to copy the entry.
